+++ packages/entity_store_firestore/lib/src/in_memory_repository/query.dart
@@ -207,4 +207,9 @@ class InMemoryRepositoryQuery<Id, E extends Entity<Id>>
   Future<Result<E?, Exception>> findOne() async {
     return (await findAll()).mapOk((ok) => ok.firstOrNull);
   }
+
+  @override
+  Future<Result<int, Exception>> count() async {
+    return (await findAll()).mapOk((ok) => ok.length);
+  }
 }