+++ packages/entity_store_firestore/lib/src/local_storage_repository/query.dart
@@ -210,4 +210,9 @@ class LocalStorageRepositoryQuery<Id, E extends Entity<Id>>
   Future<Result<E?, Exception>> findOne() async {
     return (await findAll()).mapOk((ok) => ok.firstOrNull);
   }
+
+  @override
+  Future<Result<int, Exception>> count() async {
+    return (await findAll()).mapOk((ok) => ok.length);
+  }
 }