+++ packages/entity_store_firestore/lib/src/local_storage_repository/repository.dart
@@ -84,6 +84,11 @@ abstract class LocalStorageRepository<Id, E extends Entity<Id>>
   }
 
   @override
+  Future<Result<int, Exception>> count() {
+    return query().count();
+  }
+
+  @override
   IRepositoryQuery<Id, E> query() {
     return LocalStorageRepositoryQuery(this);
   }